refactor: app folder
