+++ apps/frontend/src/app/App.vue
